pyproject.toml
[tool.poetry]
name = "mpsiemlib"
version = "1.6.0"
description = "MaxPatrol SIEM API SDK"
authors = ["nikolaiav", "GenRockeR"]
maintainers = ["GenRockeR"]
license = "GPLv3"
readme = "README.md"
homepage = "https://pypi.org/project/mpsiemlib/"
repository = "https://github.com/GenRockeR/mpsiemlib"
include = ["CHANGELOG.md"]

[tool.poetry.dependencies]
python = "^3.10"
requests = ">=2.32.3,<2.33.0"
setuptools = ">=75.6.0,<75.7.0"
pyyaml = ">=6.0.2,<6.1.0"
pytz = ">=2024.2,<2025.0"
elasticsearch = ">=7.17.12,<7.18.0"
urllib3 = ">=1.26.20,<1.27.0"
faker = ">=33.1.0,<33.2.0"
sphinx = "^8.1.3"
sphinx-rtd-theme = "^3.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
